test: add test suite for Address model
Association Tests:
it { should belong_to(:user) }
This test checks that an Address object is associated with a User object. In other words, it ensures that the Address model has a belongs_to :user association.

Validation Tests:
it { should validate_presence_of(:cep) }
it { should validate_presence_of(:street_name) }
it { should validate_presence_of(:complement) }
it { should validate_presence_of(:neighborhood) }
it { should validate_presence_of(:city) }
it { should validate_presence_of(:state) }
These tests check that the Address model has presence validations for the attributes cep, street_name, complement, neighborhood, city, and state. This means that these fields must be present for an Address object to be valid.

Uniqueness Validation Test:
describe "uniqueness validation" do ... end
This test checks that the Address model has a uniqueness validation for the cep attribute, scoped to the user_id. This means that within the scope of a single user, the cep (Postal Code) must be unique. The test also checks that the error message for this validation is as expected (specified in the I18n translation file under the key errors.duplicate_address_error).

Valid Address Test:
describe "valid address" do ... end
This test checks that an Address object with valid attributes is indeed valid. It creates a User object and an associated Address object with valid attributes (using factories, presumably defined elsewhere in the test suite), and then checks that the Address object is valid.

Invalid Address Test:
describe "invalid address" do ... end
This test checks that an Address object without a cep (Postal Code) is not valid. It creates a User object and an associated Address object with a nil cep, and then checks that the Address object is not valid.